Adds `state_header_path` argument
The factory method to uppload the bytecode is changed to also accespt
the state header for upgrade checks. This PR adds the logic to send this
state header to the factory along with the wasm.

I also changed the `path` parameter to be a positional parameter instead
of non-optional named parameter (just because having non-optional named
parameters is not cool).